refactor(app): split http transport helpers
